fix: Notification

+++ src/main/java/com/ndb/auction/models/tier/StakeTask.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 @NoArgsConstructor
 public class StakeTask {
 
-    private long expiredTime;
+    private int expiredTime;
     private float ratio;
 
 }
